fix(training_pipeline): 修复 Novely 指标中的类型错误
- 在 Novelty 类中，将 popularity_data 的默认值转换为 torch.float32 类型
- 这个修改解决了在使用 GPU 训练时出现的类型错误问题

+++ training_pipeline/metrics.py
@@ -76,7 +76,7 @@ class Novelty(PLMetric):
         """
         max_popularity = np.sum(np.sort(popularity_data)[-self.k :])
         self.add_state(
-            "popularity_data", default=torch.tensor(popularity_data / max_popularity)
+            "popularity_data", default=torch.tensor(popularity_data / max_popularity, dtype=torch.float32)
         )
 
         self.add_state(
